requirements.txt
annotated-types==0.7.0
anyio==4.9.0
asyncpg==0.30.0
bcrypt==4.1.2
cffi==1.17.1
click==8.2.1
cryptography==45.0.5
databases==0.9.0
Deprecated==1.2.18
ecdsa==0.19.1
fastapi==0.115.14
greenlet==3.2.3
h11==0.16.0
httptools==0.6.4
idna==3.10
limits==5.4.0
packaging==25.0
passlib==1.7.4
psycopg2-binary==2.9.10
pyasn1==0.6.1
pycparser==2.22
pydantic==2.11.7
pydantic_core==2.33.2
python-dotenv==1.1.1
python-jose==3.5.0
python-multipart==0.0.20
PyYAML==6.0.2
rsa==4.9.1
six==1.17.0
slowapi==0.1.9
sniffio==1.3.1
SQLAlchemy==2.0.41
starlette==0.46.2
typing-inspection==0.4.1
typing_extensions==4.14.0
uvicorn==0.35.0
uvloop==0.21.0
watchfiles==1.1.0
websockets==15.0.1
wrapt==1.17.2
